Refactor: extracted annotation related classes into their own files. Cleaned up the DataNodeEmitter
