+++ amf-webapi/shared/src/main/scala/amf/plugins/document/webapi/contexts/emitter/raml/RamlSpecEmitterContext.scala
@@ -20,6 +20,15 @@ import amf.plugins.document.webapi.contexts.{RefEmitter, SpecEmitterContext, Spe
 import amf.plugins.document.webapi.model.{Extension, Overlay}
 import amf.plugins.document.webapi.parser.RamlHeader
 import amf.plugins.document.webapi.parser.spec.declaration._
+import amf.plugins.document.webapi.parser.spec.declaration.emitters.annotations.{
+  AnnotationEmitter,
+  AnnotationTypeEmitter,
+  AnnotationsEmitter,
+  FacetsInstanceEmitter,
+  RamlAnnotationEmitter,
+  RamlAnnotationTypeEmitter,
+  RamlFacetsInstanceEmitter
+}
 import amf.plugins.document.webapi.parser.spec.declaration.emitters.raml.{
   Raml08TypePartEmitter,
   Raml10TypePartEmitter,
