+++ amf-webapi/shared/src/main/scala/amf/plugins/document/webapi/parser/spec/async/AsyncApi20DocumentEmitter.scala
@@ -14,7 +14,7 @@ import amf.plugins.document.webapi.parser.spec.async.emitters.{
   AsyncDeclarationsEmitters
 }
 import amf.plugins.document.webapi.parser.spec.common.DeclarationsEmitterWrapper
-import amf.plugins.document.webapi.parser.spec.declaration.AnnotationsEmitter
+import amf.plugins.document.webapi.parser.spec.declaration.emitters.annotations.AnnotationsEmitter
 import amf.plugins.document.webapi.parser.spec.domain.SecurityRequirementsEmitter
 import amf.plugins.document.webapi.parser.spec.oas.emitters.{InfoEmitter, TagsEmitter}
 import amf.plugins.domain.shapes.models.CreativeWork
