+++ amf-webapi/shared/src/main/scala/amf/plugins/document/webapi/parser/spec/async/emitters/AsyncSecuritySchemesEmitter.scala
@@ -4,7 +4,7 @@ import amf.core.emitter.BaseEmitters.{MapEntryEmitter, ValueEmitter, pos, traver
 import amf.core.emitter.{EntryEmitter, PartEmitter, SpecOrdering}
 import amf.core.parser.Position
 import amf.plugins.document.webapi.contexts.SpecEmitterContext
-import amf.plugins.document.webapi.parser.spec.declaration.AnnotationsEmitter
+import amf.plugins.document.webapi.parser.spec.declaration.emitters.annotations.AnnotationsEmitter
 import amf.plugins.document.webapi.parser.spec.oas.emitters.Oas3OAuth2SettingsEmitters
 import amf.plugins.domain.webapi.metamodel.security._
 import amf.plugins.domain.webapi.models.security._
