+++ amf-webapi/shared/src/main/scala/amf/plugins/document/webapi/parser/spec/common/PayloadSerializer.scala
@@ -10,7 +10,7 @@ import amf.core.remote.Payload
 import amf.core.services.RuntimeSerializer
 import amf.core.unsafe.PlatformSecrets
 import amf.core.utils._
-import amf.plugins.document.webapi.parser.spec.declaration.DataNodeEmitter
+import amf.plugins.document.webapi.parser.spec.declaration.emitters.annotations.DataNodeEmitter
 import amf.plugins.domain.shapes.models.Example
 import org.yaml.model.YDocument
