+++ amf-webapi/shared/src/main/scala/amf/plugins/document/webapi/parser/spec/declaration/emitters/SimpleTypeEmitter.scala
@@ -4,7 +4,7 @@ import amf.core.emitter.BaseEmitters.{EntryPartEmitter, MapEntryEmitter, ValueEm
 import amf.core.emitter.{EntryEmitter, SpecOrdering}
 import amf.core.metamodel.domain.ShapeModel
 import amf.plugins.document.webapi.contexts.emitter.raml.{RamlScalarEmitter, RamlSpecEmitterContext}
-import amf.plugins.document.webapi.parser.spec.declaration.DataNodeEmitter
+import amf.plugins.document.webapi.parser.spec.declaration.emitters.annotations.DataNodeEmitter
 import amf.plugins.document.webapi.parser.spec.declaration.emitters.raml.RamlCommonOASFieldsEmitter
 import amf.plugins.document.webapi.parser.spec.domain.SingleExampleEmitter
 import amf.plugins.domain.shapes.metamodel.ScalarShapeModel
