+++ amf-webapi/shared/src/main/scala/amf/plugins/document/webapi/parser/spec/domain/PayloadEmitters.scala
@@ -10,6 +10,7 @@ import amf.core.parser.{FieldEntry, Position}
 import amf.plugins.document.webapi.annotations.ParsedJSONSchema
 import amf.plugins.document.webapi.contexts.emitter.raml.RamlSpecEmitterContext
 import amf.plugins.document.webapi.parser.spec.declaration._
+import amf.plugins.document.webapi.parser.spec.declaration.emitters.annotations.AnnotationsEmitter
 import amf.plugins.document.webapi.parser.spec.declaration.emitters.raml.{
   Raml08TypePartEmitter,
   Raml10TypeEmitter,
