+++ amf-webapi/shared/src/main/scala/amf/plugins/document/webapi/parser/spec/oas/OasModuleEmitter.scala
@@ -13,6 +13,7 @@ import amf.plugins.document.webapi.contexts.emitter.oas.{JsonSchemaEmitterContex
 import amf.plugins.document.webapi.model._
 import amf.plugins.document.webapi.parser.OasHeader
 import amf.plugins.document.webapi.parser.spec.declaration._
+import amf.plugins.document.webapi.parser.spec.declaration.emitters.annotations.DataNodeEmitter
 import amf.plugins.document.webapi.parser.spec.declaration.emitters.oas
 import amf.plugins.document.webapi.parser.spec.declaration.emitters.oas.OasTypeEmitter
 import amf.plugins.document.webapi.parser.spec.domain.NamedExampleEmitter
