+++ amf-webapi/shared/src/main/scala/amf/plugins/document/webapi/parser/spec/oas/emitters/OperationEmitter.scala
@@ -11,6 +11,7 @@ import amf.plugins.document.webapi.annotations.FormBodyParameter
 import amf.plugins.document.webapi.contexts.emitter.oas.{Oas3SpecEmitterFactory, OasSpecEmitterContext}
 import amf.plugins.document.webapi.contexts.emitter.raml.Raml10SpecEmitterContext
 import amf.plugins.document.webapi.parser.spec.declaration._
+import amf.plugins.document.webapi.parser.spec.declaration.emitters.annotations.AnnotationsEmitter
 import amf.plugins.document.webapi.parser.spec.declaration.emitters.raml.{
   Raml10TypePartEmitter,
   RamlNamedTypeEmitter,
