+++ amf-webapi/shared/src/main/scala/amf/plugins/document/webapi/parser/spec/raml/emitters/RamlSecuritySchemeEmitters.scala
@@ -8,6 +8,7 @@ import amf.core.parser.{FieldEntry, Fields, Position}
 import amf.plugins.document.webapi.contexts.SpecEmitterContext
 import amf.plugins.document.webapi.contexts.emitter.raml.{RamlScalarEmitter, RamlSpecEmitterContext}
 import amf.plugins.document.webapi.parser.spec.declaration._
+import amf.plugins.document.webapi.parser.spec.declaration.emitters.annotations.{AnnotationsEmitter, DataNodeEmitter}
 import amf.plugins.document.webapi.parser.spec.declaration.emitters.raml.{Raml10TypePartEmitter, RamlNamedTypeEmitter}
 import amf.plugins.document.webapi.parser.spec.domain._
 import amf.plugins.domain.shapes.models.AnyShape
